+++ client/.env
@@ -1,0 +1,1 @@
+HOSTNAME="http://localhost:3000"
