+++ client/package.json
@@ -17,9 +17,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
     "axios": "^1.5.1",
-    "body-parser": "^1.20.2",
     "framer-motion": "^6.5.1",
-    "mongoose": "^7.5.3",
     "pill-pity": "^0.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
