+++ client/src/pages/components/footer.jsx
