+++ client/src/pages/components/frequentlyaskedquestions.jsx
