+++ client/src/pages/components/hero.jsx
