+++ client/src/pages/components/productsBrochure.jsx
