--- src/todo
@@ -1,3 +1,0 @@
-Dashboard page :
-Sidebar harusnya menu dropdown dashboard, settings, dkk
-button add
